safiri-asset-intelligence/backend/requirements.txt
# FastAPI web framework
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.3.0

# Data processing & ML
numpy==1.26.4
scikit-learn==1.3.2
pandas==2.1.1

# Database & ORM
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
neo4j==6.1.0
alembic==1.12.0

# Elasticsearch
elasticsearch==8.11.0

# Security & authentication
PyJWT==2.8.0
cryptography==41.0.3

# Fuzzy matching / string processing
rapidfuzz==3.14.3

# HTTP / async utilities
httpx==0.27.2
requests==2.32.0

# Environment & config
python-dotenv==1.0.0

# Email and SMS services
email-validator==2.0.0
twilio==8.2.2

# Optional: for logging and monitoring
loguru==0.7.2
